+++ docs/ProcessManager.Orchestrations.Abstractions/ReleaseNotes/ReleaseNotes.md
@@ -1,5 +1,11 @@
 # ProcessManager.Orchestrations.Abstractions Release Notes
 
+## Version 0.17.2
+
+- Update `RequestCalculatedWholesaleServicesInputV1` model with `IBusinessValidatedDto` interface.
+- Rename `RequestCalculatedWholesaleServicesInputV1.ChargeTypeInputV1` to `RequestCalculatedWholesaleServicesInputV1.ChargeTypeInput`.
+- Update `RequestCalculatedWholesaleServicesRejectedV1` properties to be a list of validation errors.
+
 ## Version 0.17.1
 
 - Update `RequestCalculatedEnergyTimeSeriesInputV1` with the required input.
@@ -16,7 +22,7 @@
 ## Version 0.16.0
 
 - Add `SettlementMethod` and `SettlementVersion` DataHub types.
-- Update `RequestCalculatedEnergyTimeSeries` model with `IBusinessValidatedDto` interface.
+- Update `RequestCalculatedEnergyTimeSeriesInputV1` model with `IBusinessValidatedDto` interface.
 - Update `RequestCalculatedEnergyTimeSeriesRejectedV1` properties to be a list of validation errors.
 
 ## Version 0.15.2